refactor dynamics.py; fix message for rejected simulations
