+++ src/scenic/core/simulators.py
@@ -19,7 +19,8 @@ import time
 import types
 
 from scenic.core.distributions import RejectionException
-import scenic.core.dynamics as dynamics
+from scenic.core.dynamics import GuardViolation, RejectSimulationException
+from scenic.core.dynamics.actions import Action, _EndScenarioAction, _EndSimulationAction
 import scenic.core.errors as errors
 from scenic.core.errors import InvalidScenarioError, optionallyDebugRejection
 from scenic.core.object_types import (
@@ -50,12 +51,6 @@ class SimulationCreationError(Exception):
 
 class DivergenceError(Exception):
     """Exception indicating simulation replay failed due to simulator nondeterminism."""
-
-    pass
-
-
-class RejectSimulationException(Exception):
-    """Exception indicating a requirement was violated at runtime."""
 
     pass
 
@@ -219,17 +214,13 @@ class Simulator(abc.ABC):
                 verbosity=verbosity,
                 **kwargs,
             )
-        except (
-            RejectSimulationException,
-            RejectionException,
-            dynamics.GuardViolation,
-        ) as e:
+        except (RejectSimulationException, RejectionException, GuardViolation) as e:
             if verbosity >= 2:
                 print(
                     f"  Rejected simulation {name} at time step "
-                    f"{simulation.currentTime} because: {e}"
+                    f"{e.simulation.currentTime} because: {e}"
                 )
-            if raiseGuardViolations and isinstance(e, dynamics.GuardViolation):
+            if raiseGuardViolations and isinstance(e, GuardViolation):
                 raise
             else:
                 optionallyDebugRejection(e)
@@ -397,7 +388,11 @@ class Simulation(abc.ABC):
                 self.records,
             )
             self.result = result
-            return self
+        except (RejectSimulationException, RejectionException, GuardViolation) as e:
+            # This simulation will be thrown out, but attach it to the exception
+            # to aid in debugging.
+            e.simulation = self
+            raise
         finally:
             self.destroy()
             for obj in self.objects:
@@ -454,9 +449,9 @@ class Simulation(abc.ABC):
                 actions = agent.behavior._step()
 
                 # Handle pseudo-actions marking the end of a simulation/scenario
-                if isinstance(actions, EndSimulationAction):
+                if isinstance(actions, _EndSimulationAction):
                     return TerminationType.terminatedByBehavior, str(actions)
-                elif isinstance(actions, EndScenarioAction):
+                elif isinstance(actions, _EndScenarioAction):
                     scenario = actions.scenario
                     if scenario._isRunning:
                         scenario._stop(actions)
@@ -847,68 +842,6 @@ class DummySimulation(Simulation):
             if prop not in vals:
                 vals[prop] = None
         return vals
-
-
-class Action(abc.ABC):
-    """An :term:`action` which can be taken by an agent for one step of a simulation."""
-
-    def canBeTakenBy(self, agent):
-        """Whether this action is allowed to be taken by the given agent.
-
-        The default implementation always returns True.
-        """
-        return True
-
-    @abc.abstractmethod
-    def applyTo(self, agent, simulation):
-        """Apply this action to the given agent in the given simulation.
-
-        This method should call simulator APIs so that the agent will take this action
-        during the next simulated time step. Depending on the simulator API, it may be
-        necessary to batch each agent's actions into a single update: in that case you
-        can have this method set some state on the agent, then apply the actual update
-        in an overridden implementation of `Simulation.executeActions`. For examples,
-        see the CARLA interface: `scenic.simulators.carla.actions` has some CARLA-specific
-        actions which directly call CARLA APIs, while the generic steering and braking
-        actions from `scenic.domains.driving.actions` are implemented using the batching
-        approach (see for example the ``setThrottle`` method of the class
-        `scenic.simulators.carla.model.Vehicle`, which sets state later read by
-        ``CarlaSimulation.executeActions`` in `scenic.simulators.carla.simulator`).
-        """
-        raise NotImplementedError
-
-
-class EndSimulationAction(Action):
-    """Special action indicating it is time to end the simulation.
-
-    Only for internal use.
-    """
-
-    def __init__(self, line):
-        self.line = line
-
-    def __str__(self):
-        return f'"terminate simulation" executed on line {self.line}'
-
-    def applyTo(self, agent, simulation):
-        assert False
-
-
-class EndScenarioAction(Action):
-    """Special action indicating it is time to end the current scenario.
-
-    Only for internal use.
-    """
-
-    def __init__(self, scenario, line):
-        self.scenario = scenario
-        self.line = line
-
-    def __str__(self):
-        return f'"terminate" executed on line {self.line}'
-
-    def applyTo(self, agent, simulation):
-        assert False
 
 
 @enum.unique
